We removed those integration tests because they are quite expensive
lasting for almost an hour and we can test that in the unit level by
checking if the errands execute the expected cf commands in order to
enable a service and register the broker. We rely on other tests in our
pipeline to fail if there's any integration issue when executing bosh
run errand.

[#142052159]

Signed-off-by: Alberto Leal <aleal@pivotal.io>