Refactor some ssh_gateway tests
Some ssh_gateway tests were being used to check whether it is possible
or not to use service keys provived by CF to access a service when the
service key does not exist anymore.

We simplified those tests by just checking that the information present in
the service key is there (which is provided by the multitenant broker).
This is not a system test anymore. The system test is being played by
the bind/unbind tests, which call the same function inside the broker.

[#149735781]

Signed-off-by: Vlad Stoian <vstoian@pivotal.io>