Fix non-TLS AMQP scheme when TLS enabled
As of https://github.com/pivotal-cf/cf-rabbitmq-release/pull/394, a bug
was fixed which disabled the non-TLS AMQP listener when TLS was enabled
& disable_non_ssl_listeners was disabled.

This revealed a bug in the multitenant broker which assumed that only
one of AMQP or AMQPS can be enabled at a time, resulting in the
incorrect scheme for non-TLS AMQP bindings. This PR fixes that second
bug.