Force use of IP addresses from BOSH DNS in CI.
For the same reason as the failures in cf-rabbitmq-release CI, the CF
Deployment environments used for testing now use BOSH DNS Addresses by
default. This causes the erl_inetrc file to be mispopulated with
hostnames instead of IP adddresses, leading to a failure in the
pre-start script. When deploying this release in CI, we now set the
consumes block of the cf-rabbitm-release rabbitmq-server job to have
rabbitmq-server.ip_address true.